Use TSLint and Prettier to lint and format code
Assuming the VSCode TSLint extension is installed auto-fixable lint
errors will be fixed whenever a document is manually saved. It's
also possible to run the linter manually via `npm run lint`, but
that won't auto-fix anything.

Prettier has to be run manually for now via `npm run format`, would
be nice to get it to run on-save in VSCode, but it must run after
TSLint and I don't know how to specify the order in which extensions
are executed.