Don't allow the user to add directories that aren't Git repositories
The path entered by the user on the `Add Local Repository` dialog is
now checked to ensure it's actually the root of a git repository,
and if it turns out that it's not the dialog will remain open. This
is obviously pretty poor UX, the dialog should display the
validation error so the user knows what's going on.